Nonlinear Modeling

More information can be found in the Nonlinear section of the manual.

JuMP.Nonlinear.NonlinearData — Type

NonlinearData()

Construct a new MOI.AbstractNLPEvaluator object.

Expressions

JuMP.Nonlinear.ExpressionIndex — Type

ExpressionIndex

JuMP.Nonlinear.add_expression — Function

add_expression(data::NonlinearData, expr)

Parameters

JuMP.Nonlinear.ParameterIndex — Type

ParameterIndex

JuMP.Nonlinear.add_parameter — Function

add_parameter(data::NonlinearData, value::Float64)

JuMP.Nonlinear.set_parameter — Function

set_parameter(data::NonlinearData, p::ParameterIndex, value::Float64)

Objectives

JuMP.Nonlinear.set_objective — Function

set_objective(data::NonlinearData, obj)

Constraints

JuMP.Nonlinear.ConstraintIndex — Type

ConstraintIndex

JuMP.Nonlinear.add_constraint — Function

add_constraint(data::NonlinearData, expr::Expr)

JuMP.Nonlinear.delete — Function

delete(data::NonlinearData, c::ConstraintIndex)

User-defined operators

JuMP.Nonlinear.register_operator — Function

register_operator(
    data::NonlinearData,
    op::Symbol,
    nargs::Int,
    f::Function,
)

register_operator(
    data::NonlinearData,
    op::Symbol,
    nargs::Int,
    f::Function,
    ∇f::Function,
)

register_operator(
    data::NonlinearData,
    op::Symbol,
    nargs::Int,
    f::Function,
    ∇f::Function,
    ∇²f::Function,
)

Automatic-differentiation backends

JuMP.Nonlinear.AbstractAutomaticDifferentiation — Type

AbstractAutomaticDifferentiation

An abstract type for extending

JuMP.Nonlinear.Default — Type

Default() <: AbstractAutomaticDifferentiation

The default implementation of AbstractAutomaticDifferentiation. This only supported :ExprGraph.

JuMP.Nonlinear.set_differentiation_backend — Function

set_differentiation_backend(
    data::NonlinearData,
    backend::AbstractAutomaticDifferentiation,
    ordered_variables::Vector{MOI.VariableIndex}
)

Set the automatic differentiationn backend of data to backend.

ordered_variables is an ordered vector containing all of the variables that are present in the model. This order corresponds to the order of the primal solution vector x that is passed to the various functions in MOI's nonlinear API.