Variables

Variables

What is a JuMP VariableRef?

DRAFT: A JuMP variable is a reference to an index in a model. It's a thin wrapper around MOI.VariableIndex. Variables have (1) names, and (2) attributes. Describe the different scopes of a variable (e.g., as Julia variables, lookup by string name, and lookup by symbol).

The @variable macro

DRAFT: Describe the complete syntax of the @variable macro. Anonymous versus named variables. Describe the three possible container types returned and how to use them (Array, JuMPArray, and Dict).

JuMP.@variable — Macro.

@variable(model, kwargs...)

Add an anonymous (see Names) variable to the model model described by the keyword arguments kwargs and returns the variable.

@variable(model, expr, args..., kwargs...)

Add a variable to the model model described by the expression expr, the positional arguments args and the keyword arguments kwargs. The expression expr can either be (note that in the following the symbol <= can be used instead of ≤ and the symbol >=can be used instead of ≥)

of the form varexpr creating variables described by varexpr;

of the form varexpr ≤ ub (resp. varexpr ≥ lb) creating variables described by varexpr with upper bounds given by ub (resp. lower bounds given by lb);

of the form varexpr == value creating variables described by varexpr with fixed values given by value; or

of the form lb ≤ varexpr ≤ ub or ub ≥ varexpr ≥ lb creating variables described by varexpr with lower bounds given by lb and upper bounds given by ub.

The expression varexpr can either be

of the form varname creating a scalar real variable of name varname;

of the form varname[...] or [...] creating a container of variables (see Containers in macro.

The recognized positional arguments in args are the following:

Bin: Sets the variable to be binary, i.e. either 0 or 1.

Int: Sets the variable to be integer, i.e. one of ..., -2, -1, 0, 1, 2, ...

Symmetric: Only available when creating a square matrix of variables, i.e. when varexpr is of the form varname[1:n,1:n] or varname[i=1:n,j=1:n]. It creates a symmetric matrix of variable, that is, it only creates a new variable for varname[i,j] with i ≤ j and sets varname[j,i] to the same variable as varname[i,j].

PSD: The square matrix of variable is both Symmetric and constrained to be positive semidefinite.

The recognized keyword arguments in kwargs are the following:

basename: Sets the base name used to generate variable names. It corresponds to the variable name for scalar variable, otherwise, the variable names are basename[...] for each indices ... of the axes axes.

lowerbound: Sets the value of the variable lower bound.

upperbound: Sets the value of the variable upper bound.

start: Sets the variable starting value used as initial guess in optimization.

binary: Sets whether the variable is binary or not.

integer: Sets whether the variable is integer or not.

variabletype: See the "Note for extending the variable macro" section below.

container: Specify the container type, see Containers in macro.

Examples

The following are equivalent ways of creating a variable x of name x with lowerbound 0:

# Specify everything in `expr`
@variable(model, x >= 0)
# Specify the lower bound using a keyword argument
@variable(model, x, lowerbound=0)
# Specify everything in `kwargs`
x = @variable(model, basename="x", lowerbound=0)

The following are equivalent ways of creating a JuMPArray of index set [:a, :b] and with respective upper bounds 2 and 3 and names x[a] and `x[b].

ub = Dict(:a => 2, :b => 3)
# Specify everything in `expr`
@variable(model, x[i=keys(ub)] <= ub[i])
# Specify the upper bound using a keyword argument
@variable(model, x[i=keys(ub)], upperbound=ub[i])

Note for extending the variable macro

The single scalar variable or each scalar variable of the container are created using addvariable(model, buildvariable(_error, info, extra_args...; extra_kwargs...)) where

model is the model passed to the @variable macro;

_error is an error function with a single String argument showing the @variable call in addition to the error message given as argument;

info is the VariableInfo struct containing the information gathered in expr, the recognized keyword arguments (except basename and variabletype) and the recognized positional arguments (except Symmetric and PSD);

extra_args are the unrecognized positional arguments of args plus the value of the variabletype keyword argument if present. The variabletype keyword argument allows the user to pass a position argument to buildvariable without the need to give a positional argument to @variable. In particular, this allows the user to give a positional argument to the buildvariable call when using the anonymous single variable syntax @variable(model, kwargs...); and

extra_kwargs are the unrecognized keyword argument of kwargs.

Examples

The following creates a variable x of name x with lowerbound 0 as with the first example above but does it without using the @variable macro

info = VariableInfo(true, 0, false, NaN, false, NaN, false, NaN, false, false)
JuMP.addvariable(model, JuMP.buildvariable(error, info), "x")

The following creates a JuMPArray of index set [:a, :b] and with respective upper bounds 2 and 3 and names x[a] and x[b] as with the second example above but does it without using the @variable macro

# Without the `@variable` macro
data = Vector{JuMP.variabletype(model)}(undef, length(keys(ub)))
x = JuMPArray(data, keys(ub))
for i in keys(ub)
    info = VariableInfo(false, NaN, true, ub[i], false, NaN, false, NaN, false, false)
    x[i] = JuMP.addvariable(model, JuMP.buildvariable(error, info), "x[$i]")
end

The following are equivalent ways of creating a Matrix of size N x N with variables custom variables created with a JuMP extension using the Poly(X) positional argument to specify its variables:

# Using the `@variable` macro
@variable(model, x[1:N,1:N], Symmetric, Poly(X))
# Without the `@variable` macro
x = Matrix{JuMP.variabletype(model, Poly(X))}(N, N)
info = VariableInfo(false, NaN, false, NaN, false, NaN, false, NaN, false, false)
for i in 1:N, j in i:N
    x[i,j] = x[j,i] = JuMP.addvariable(model, buildvariable(error, info, Poly(X)), "x[$i,$j]")
end

source

How to delete variables.