Extensions

Extending JuMP

TODO: How to extend JuMP: discussion on different ways to build on top of JuMP. How to extend JuMP's macros and how to avoid doing this.

Extending MOI

TODO: Create new MOI function/sets, how to use it in JuMP, create new bridge

Extending JuMP macros

In order to provide a convenient syntax for the user to create variables, constraints or set the objective of a JuMP extension, it might be required to use macros similar to @variable, @constraint and @objective. It is recommended to first check whether it is possible to extend one of these three macros before creating a new one so as to leverage all their features and provide a more consistent interface to the user.

Extending the @variable macro

TODO: parse/build/add

Extending the @constraint macro

The @constraint macro always calls the same three functions:

parse_constraint: is called at parsing time, it parses the constraint expression and returns a build_constraint call expression;

build_constraint: given the functions and sets involved in the constraints, it returns a AbstractConstraint;

add_constraint: given the model, the AbstractConstraint constructed in build_constraint and the constraint name, it stores them in the model and returns a ConstraintRef.

Adding methods to these functions is the recommended way to extend the @constraint macro.

Adding build_constraint methods

There is typically two choices when creating a build_constraint method, either return an AbstractConstraint already supported by the model, i.e. ScalarConstraint or VectorConstraint, or a custom AbstractConstraint with a corresponding add_constraint method (see Adding add_constraint methods).

JuMP.build_constraint — Function.

function build_constraint(_error::Function, Q::Symmetric{V, M},
                          ::PSDCone) where {V <: AbstractJuMPScalar,
                                            M <: AbstractMatrix{V}}

Return a VectorConstraint of shape SymmetricMatrixShape constraining the matrix Q to be positive semidefinite.

This function is used by the @variable macro to create a symmetric semidefinite matrix of variables and by the @constraint macros as follows:

@constraint(model, Symmetric(Q) in PSDCone())

The form above is usually used when the entries of Q are affine or quadratic expressions but it can also be used when the entries are variables to get the reference of the semidefinite constraint, e.g.,

@variable model Q[1:2,1:2] Symmetric
# The type of `Q` is `Symmetric{VariableRef, Matrix{VariableRef}}`
var_psd = @constraint model Q in PSDCone()
# The `var_psd` variable contains a reference to the constraint

function build_constraint(_error::Function,
                          Q::AbstractMatrix{<:AbstractJuMPScalar},
                          ::PSDCone)

Return a VectorConstraint of shape SquareMatrixShape constraining the matrix Q to be symmetric and positive semidefinite.

This function is used by the @constraint and @SDconstraint macros as follows:

@constraint(model, Q in PSDCone())
@SDconstraint(model, P ⪰ Q)

The @constraint call above is usually used when the entries of Q are affine or quadratic expressions but it can also be used when the entries are variables to get the reference of the semidefinite constraint, e.g.,

@variable model Q[1:2,1:2]
# The type of `Q` is `Matrix{VariableRef}`
var_psd = @constraint model Q in PSDCone()
# The `var_psd` variable contains a reference to the constraint

Shapes

Shapes allow vector constraints, which are represented as flat vectors in MOI, to retain a matrix shape at the JuMP level. There is a shape field in VectorConstraint that can be set in build_constraint and that is used to reshape the result computed in value and dual.

JuMP.AbstractShape — Type.

AbstractShape

Abstract vectorizable shape. Given a flat vector form of an object of shape shape, the original object can be obtained by reshape.

JuMP.shape — Function.

shape(c::AbstractConstraint)::AbstractShape

Return the shape of the constraint c.

JuMP.reshape — Function.

reshape(vectorized_form::Vector, shape::AbstractShape)

Return an object in its original shape shape given its vectorized form vectorized_form.

Examples

Given a SymmetricMatrixShape of vectorized form [1, 2, 3], the following code returns the matrix Symmetric(Matrix[1 2; 2 3]):

reshape([1, 2, 3], SymmetricMatrixShape(2))

JuMP.dual_shape — Function.

dual_shape(shape::AbstractShape)::AbstractShape

Returns the shape of the dual space of the space of objects of shape shape. By default, the dual_shape of a shape is itself. See the examples section below for an example for which this is not the case.

Examples

Consider polynomial constraints for which the dual is moment constraints and moment constraints for which the dual is polynomial constraints. Shapes for polynomials can be defined as follows:

struct Polynomial
    coefficients::Vector{Float64}
    monomials::Vector{Monomial}
end
struct PolynomialShape <: JuMP.AbstractShape
    monomials::Vector{Monomial}
end
JuMP.reshape(x::Vector, shape::PolynomialShape) = Polynomial(x, shape.monomials)

and a shape for moments can be defined as follows:

struct Moments
    coefficients::Vector{Float64}
    monomials::Vector{Monomial}
end
struct MomentsShape <: JuMP.AbstractShape
    monomials::Vector{Monomial}
end
JuMP.reshape(x::Vector, shape::MomentsShape) = Moments(x, shape.monomials)

The dual_shape allows to define the shape of the dual of polynomial and moment constraints:

dual_shape(shape::PolynomialShape) = MomentsShape(shape.monomials)
dual_shape(shape::MomentsShape) = PolynomialShape(shape.monomials)

JuMP.ScalarShape — Type.

ScalarShape

Shape of scalar constraints.

JuMP.VectorShape — Type.

VectorShape

Vector for which the vectorized form corresponds exactly to the vector given.

JuMP.SquareMatrixShape — Type.

SquareMatrixShape

Shape object for a square matrix of side_dimension rows and columns. The vectorized form contains the entries of the the matrix given column by column (or equivalently, the entries of the lower-left triangular part given row by row).

JuMP.SymmetricMatrixShape — Type.

SymmetricMatrixShape

Shape object for a symmetric square matrix of side_dimension rows and columns. The vectorized form contains the entries of the upper-right triangular part of the matrix given column by column (or equivalently, the entries of the lower-left triangular part given row by row).

Adding add_constraint methods

TODO

JuMP.add_constraint — Function.

add_constraint(model::Model, c::AbstractConstraint, name::String="")

Add a constraint c to Model model and sets its name.

Adding parse_constraint methods

TODO

Extending the @objective macro

TODO

Defining new JuMP models

TODO