Bridge interface

A bridge should implement the following functions to be usable by a bridge optimizer:

MathOptInterface.Bridges.added_constrained_variable_types — Function

added_constrained_variable_types(BT::Type{<:Variable.AbstractBridge})::Vector{Tuple{DataType}}

Return a list of the types of constrained variables that bridges of concrete type BT add. This is used by the LazyBridgeOptimizer.

MathOptInterface.Bridges.added_constraint_types — Function

added_constraint_types(BT::Type{<:Constraint.AbstractBridge})::Vector{Tuple{DataType, DataType}}

Return a list of the types of constraints that bridges of concrete type BT add. This is used by the LazyBridgeOptimizer.

Additionally, variable bridges should implement:

MathOptInterface.Bridges.Variable.supports_constrained_variable — Function

supports_constrained_variable(::Type{<:AbstractBridge},
                               ::Type{<:MOI.AbstractSet})::Bool

Return a Bool indicating whether the bridges of type BT support bridging constrained variables in S.

MathOptInterface.Bridges.Variable.concrete_bridge_type — Function

concrete_bridge_type(
    BT::Type{<:AbstractBridge},
    S::Type{<:MOI.AbstractSet},
)::DataType

Return the concrete type of the bridge supporting variables in S constraints. This function can only be called if MOI.supports_constrained_variable(BT, S) is true.

Examples

As a variable in MathOptInterface.GreaterThan is bridged into variables in MathOptInterface.Nonnegatives by the VectorizeBridge:

MOI.Bridges.Variable.concrete_bridge_type(
    MOI.Bridges.Variable.VectorizeBridge{Float64},
    MOI.GreaterThan{Float64},
)

# output

MathOptInterface.Bridges.Variable.VectorizeBridge{Float64,MathOptInterface.Nonnegatives}

MathOptInterface.Bridges.Variable.bridge_constrained_variable — Function

bridge_constrained_variable(BT::Type{<:AbstractBridge}, model::MOI.ModelLike,
                            set::MOI.AbstractSet)

Bridge the constrained variable in set using bridge BT to model and returns a bridge object of type BT. The bridge type BT should be a concrete type, that is, all the type parameters of the bridge should be set. Use concrete_bridge_type to obtain a concrete type for given set types.

constraint bridges should implement:

MathOptInterface.supports_constraint — Method

MOI.supports_constraint(BT::Type{<:AbstractBridge}, F::Type{<:MOI.AbstractFunction}, S::Type{<:MOI.AbstractSet})::Bool

Return a Bool indicating whether the bridges of type BT support bridging F-in-S constraints.

MathOptInterface.Bridges.Constraint.concrete_bridge_type — Function

concrete_bridge_type(
    BT::Type{<:AbstractBridge},
    F::Type{<:MOI.AbstractFunction},
    S::Type{<:MOI.AbstractSet}
)::DataType

Return the concrete type of the bridge supporting F-in-S constraints. This function can only be called if MOI.supports_constraint(BT, F, S) is true.

Examples

As a MathOptInterface.SingleVariable-in-MathOptInterface.Interval constraint is bridged into a MathOptInterface.SingleVariable-in-MathOptInterface.GreaterThan and a MathOptInterface.SingleVariable-in-MathOptInterface.LessThan by the SplitIntervalBridge:

MOI.Bridges.Constraint.concrete_bridge_type(
    MOI.Bridges.Constraint.SplitIntervalBridge{Float64},
    MOI.SingleVariable,
    MOI.Interval{Float64},
)

# output

MathOptInterface.Bridges.Constraint.SplitIntervalBridge{Float64,MathOptInterface.SingleVariable,MathOptInterface.Interval{Float64},MathOptInterface.GreaterThan{Float64},MathOptInterface.LessThan{Float64}}

MathOptInterface.Bridges.Constraint.bridge_constraint — Function

bridge_constraint(BT::Type{<:AbstractBridge}, model::MOI.ModelLike,
                  func::AbstractFunction, set::MOI.AbstractSet)

Bridge the constraint func-in-set using bridge BT to model and returns a bridge object of type BT. The bridge type BT should be a concrete type, that is, all the type parameters of the bridge should be set. Use concrete_bridge_type to obtain a concrete type for given function and set types.

and objective bridges should implement:

MathOptInterface.Bridges.set_objective_function_type — Function

set_objective_function_type(BT::Type{<:Objective.AbstractBridge})::Type{<:MOI.AbstractScalarFunction}

Return the type of objective function that bridges of concrete type BT set. This is used by the LazyBridgeOptimizer.

MathOptInterface.Bridges.Objective.concrete_bridge_type — Function

concrete_bridge_type(BT::Type{<:MOI.Bridges.Objective.AbstractBridge},
                     F::Type{<:MOI.AbstractScalarFunction})::DataType

Return the concrete type of the bridge supporting objective functions of type F. This function can only be called if MOI.supports_objective_function(BT, F) is true.

MathOptInterface.Bridges.Objective.bridge_objective — Function

bridge_objective(BT::Type{<:MOI.Bridges.Objective.AbstractBridge},
                 model::MOI.ModelLike,
                 func::MOI.AbstractScalarFunction)

Bridge the objective function func using bridge BT to model and returns a bridge object of type BT. The bridge type BT should be a concrete type, that is, all the type parameters of the bridge should be set. Use concrete_bridge_type to obtain a concrete type for a given function type.

When querying the NumberOfVariables, NumberOfConstraints and ListOfConstraintIndices, the variables and constraints created by the bridges in the underlying model are hidden by the bridge optimizer. For this purpose, the bridge should provide access to the variables and constraints it has creates by implemented the following methods of get:

MathOptInterface.get — Method

MOI.get(b::AbstractBridge, ::MOI.NumberOfVariables)

The number of variables created by the bridge b in the model.

MathOptInterface.get — Method

MOI.get(b::AbstractBridge, ::MOI.NumberOfVariables)

The list of variables created by the bridge b in the model.

MathOptInterface.get — Method

MOI.get(b::AbstractBridge, ::MOI.NumberOfConstraints{F, S}) where {F, S}

The number of constraints of the type F-in-S created by the bridge b in the model.

MathOptInterface.get — Method

MOI.get(b::AbstractBridge, ::MOI.ListOfConstraintIndices{F, S}) where {F, S}

A Vector{ConstraintIndex{F,S}} with indices of all constraints of type F-inS created by the bride b in the model (i.e., of length equal to the value of NumberOfConstraints{F,S}()).