Info

This documentation is a copy of the official MathOptInterface documentation available at https://jump.dev/MathOptInterface.jl/v0.9.21. It is included here to make it easier to link concepts between JuMP and MathOptInterface.

Utilities.Model

MathOptInterface.Utilities.Model — Type

An implementation of ModelLike that supports all functions and sets defined in MOI. It is parameterized by the coefficient type.

Examples

model = Model{Float64}()
x = add_variable(model)

Utilities.UniversalFallback

MathOptInterface.Utilities.UniversalFallback — Type

UniversalFallback

The UniversalFallback can be applied on a MathOptInterface.ModelLike model to create the model UniversalFallback(model) supporting any constraint and attribute. This allows to have a specialized implementation in model for performance critical constraints and attributes while still supporting other attributes with a small performance penalty. Note that model is unaware of constraints and attributes stored by UniversalFallback so this is not appropriate if model is an optimizer (for this reason, MathOptInterface.optimize! has not been implemented). In that case, optimizer bridges should be used instead.

Utilities.@macro

MathOptInterface.Utilities.@model — Macro

macro model(
    model_name,
    scalar_sets,
    typed_scalar_sets,
    vector_sets,
    typed_vector_sets,
    scalar_functions,
    typed_scalar_functions,
    vector_functions,
    typed_vector_functions,
    is_optimizer = false
)

Creates a type model_name implementing the MOI model interface and containing scalar_sets scalar sets typed_scalar_sets typed scalar sets, vector_sets vector sets, typed_vector_sets typed vector sets, scalar_functions scalar functions, typed_scalar_functions typed scalar functions, vector_functions vector functions and typed_vector_functions typed vector functions. To give no set/function, write (), to give one set S, write (S,).

The function MathOptInterface.SingleVariable should not be given in scalar_functions. The model supports MathOptInterface.SingleVariable-in-S constraints where S is MathOptInterface.EqualTo, MathOptInterface.GreaterThan, MathOptInterface.LessThan, MathOptInterface.Interval, MathOptInterface.Integer, MathOptInterface.ZeroOne, MathOptInterface.Semicontinuous or MathOptInterface.Semiinteger. The sets supported with the MathOptInterface.SingleVariable cannot be controlled from the macro, use the UniversalFallback to support more sets.

This macro creates a model specialized for specific types of constraint, by defining specialized structures and methods. To create a model that, in addition to be optimized for specific constraints, also support arbitrary constraints and attributes, use UniversalFallback.

If is_optimizer = true, the resulting struct is a of GenericOptimizer, which is a subtype of MathOptInterface.AbstractOptimizer, otherwise, it is a GenericModel, which is a subtype of MathOptInterface.ModelLike.

Examples

The model describing an linear program would be:

@model(LPModel,                                                   # Name of model
      (),                                                         # untyped scalar sets
      (MOI.EqualTo, MOI.GreaterThan, MOI.LessThan, MOI.Interval), #   typed scalar sets
      (MOI.Zeros, MOI.Nonnegatives, MOI.Nonpositives),            # untyped vector sets
      (),                                                         #   typed vector sets
      (),                                                         # untyped scalar functions
      (MOI.ScalarAffineFunction,),                                #   typed scalar functions
      (MOI.VectorOfVariables,),                                   # untyped vector functions
      (MOI.VectorAffineFunction,),                                #   typed vector functions
      false
    )

Let MOI denote MathOptInterface, MOIU denote MOI.Utilities. The macro would create the following types with struct_of_constraint_code:

struct LPModelScalarConstraints{T, C1, C2, C3, C4} <: MOIU.StructOfConstraints
    moi_equalto::C1
    moi_greaterthan::C2
    moi_lessthan::C3
    moi_interval::C4
end
struct LPModelVectorConstraints{T, C1, C2, C3} <: MOIU.StructOfConstraints
    moi_zeros::C1
    moi_nonnegatives::C2
    moi_nonpositives::C3
end
struct LPModelFunctionConstraints{T} <: MOIU.StructOfConstraints
    moi_scalaraffinefunction::LPModelScalarConstraints{
        T,
        MOIU.VectorOfConstraints{MOI.ScalarAffineFunction{T}, MOI.EqualTo{T}},
        MOIU.VectorOfConstraints{MOI.ScalarAffineFunction{T}, MOI.GreaterThan{T}},
        MOIU.VectorOfConstraints{MOI.ScalarAffineFunction{T}, MOI.LessThan{T}},
        MOIU.VectorOfConstraints{MOI.ScalarAffineFunction{T}, MOI.Interval{T}}
    }
    moi_vectorofvariables::LPModelVectorConstraints{
        T,
        MOIU.VectorOfConstraints{MOI.VectorOfVariables, MOI.Zeros},
        MOIU.VectorOfConstraints{MOI.VectorOfVariables, MOI.Nonnegatives},
        MOIU.VectorOfConstraints{MOI.VectorOfVariables, MOI.Nonpositives}
    }
    moi_vectoraffinefunction::LPModelVectorConstraints{
        T,
        MOIU.VectorOfConstraints{MOI.VectorAffineFunction{T}, MOI.Zeros},
        MOIU.VectorOfConstraints{MOI.VectorAffineFunction{T}, MOI.Nonnegatives},
        MOIU.VectorOfConstraints{MOI.VectorAffineFunction{T}, MOI.Nonpositives}
    }
end
const LPModel{T} = MOIU.GenericModel{T, LPModelFunctionConstraints{T}}

The type LPModel implements the MathOptInterface API except methods specific to optimizers like optimize! or get with VariablePrimal.

MathOptInterface.Utilities.GenericModel — Type

mutable struct GenericModel{T,C} <: AbstractModelLike{T}

Implements a models supporting

an objective function of type MOI.SingleVariable, MOI.ScalarAffineFunction{T} and MOI.ScalarQuadraticFunction{T},

MathOptInterface.SingleVariable-in-S constraints where S is MathOptInterface.EqualTo, MathOptInterface.GreaterThan, MathOptInterface.LessThan, MathOptInterface.Interval, MathOptInterface.Integer, MathOptInterface.ZeroOne, MathOptInterface.Semicontinuous or MathOptInterface.Semiinteger.

F-in-S constraints that are supported by C.

MathOptInterface.Utilities.GenericOptimizer — Type

mutable struct GenericOptimizer{T,C} <: AbstractOptimizer{T}

Implements a models supporting

an objective function of type MOI.SingleVariable, MOI.ScalarAffineFunction{T} and MOI.ScalarQuadraticFunction{T},

MathOptInterface.SingleVariable-in-S constraints where S is MathOptInterface.EqualTo, MathOptInterface.GreaterThan, MathOptInterface.LessThan, MathOptInterface.Interval, MathOptInterface.Integer, MathOptInterface.ZeroOne, MathOptInterface.Semicontinuous or MathOptInterface.Semiinteger.

F-in-S constraints that are supported by C.

MathOptInterface.Utilities.struct_of_constraint_code — Function

struct_of_constraint_code(struct_name, types, field_types = nothing)

Given a vector of n SymbolFun or SymbolSet in types, creates a struct of name struct_name that is a subtype of StructOfConstraint{T, C1, C2, ..., Cn} if field_types is nothing and a subtype of StructOfConstraint{T} otherwise. It contains n field where the ith field has type Ci if field_types is nothing and type field_types[i] otherwise. If types is vector of SymbolFun (resp. SymbolSet) then the constraints of that function (resp. set) type are stored in the corresponding field.

Caching optimizer

MathOptInterface.Utilities.CachingOptimizer — Type

CachingOptimizer

CachingOptimizer is an intermediate layer that stores a cache of the model and links it with an optimizer. It supports incremental model construction and modification even when the optimizer doesn't.

A CachingOptimizer may be in one of three possible states (CachingOptimizerState):

NO_OPTIMIZER: The CachingOptimizer does not have any optimizer.

EMPTY_OPTIMIZER: The CachingOptimizer has an empty optimizer. The optimizer is not synchronized with the cached model.

ATTACHED_OPTIMIZER: The CachingOptimizer has an optimizer, and it is synchronized with the cached model.

A CachingOptimizer has two modes of operation (CachingOptimizerMode):

MANUAL: The only methods that change the state of the CachingOptimizer are Utilities.reset_optimizer, Utilities.drop_optimizer, and Utilities.attach_optimizer. Attempting to perform an operation in the incorrect state results in an error.

AUTOMATIC: The CachingOptimizer changes its state when necessary. For example, optimize! will automatically call attach_optimizer (an optimizer must have been previously set). Attempting to add a constraint or perform a modification not supported by the optimizer results in a drop to EMPTY_OPTIMIZER mode.

MathOptInterface.Utilities.attach_optimizer — Function

attach_optimizer(model::CachingOptimizer)

Attaches the optimizer to model, copying all model data into it. Can be called only from the EMPTY_OPTIMIZER state. If the copy succeeds, the CachingOptimizer will be in state ATTACHED_OPTIMIZER after the call, otherwise an error is thrown; see MathOptInterface.copy_to for more details on which errors can be thrown.

MOIU.attach_optimizer(model::Model)

Call MOIU.attach_optimizer on the backend of model.

Cannot be called in direct mode.

MathOptInterface.Utilities.reset_optimizer — Function

reset_optimizer(m::CachingOptimizer, optimizer::MOI.AbstractOptimizer)

Sets or resets m to have the given empty optimizer optimizer.

Can be called from any state. An assertion error will be thrown if optimizer is not empty.

The CachingOptimizer m will be in state EMPTY_OPTIMIZER after the call.

reset_optimizer(m::CachingOptimizer)

Detaches and empties the current optimizer. Can be called from ATTACHED_OPTIMIZER or EMPTY_OPTIMIZER state. The CachingOptimizer will be in state EMPTY_OPTIMIZER after the call.

MOIU.reset_optimizer(model::Model, optimizer::MOI.AbstractOptimizer)

Call MOIU.reset_optimizer on the backend of model.

Cannot be called in direct mode.

MOIU.reset_optimizer(model::Model)

Call MOIU.reset_optimizer on the backend of model.

Cannot be called in direct mode.

MathOptInterface.Utilities.drop_optimizer — Function

drop_optimizer(m::CachingOptimizer)

Drops the optimizer, if one is present. Can be called from any state. The CachingOptimizer will be in state NO_OPTIMIZER after the call.

MOIU.drop_optimizer(model::Model)

Call MOIU.drop_optimizer on the backend of model.

Cannot be called in direct mode.

MathOptInterface.Utilities.state — Function

state(m::CachingOptimizer)::CachingOptimizerState

Returns the state of the CachingOptimizer m. See Utilities.CachingOptimizer.

MathOptInterface.Utilities.mode — Function

mode(m::CachingOptimizer)::CachingOptimizerMode

Returns the operating mode of the CachingOptimizer m. See Utilities.CachingOptimizer.

Copy utilities

MathOptInterface.Utilities.automatic_copy_to — Function

automatic_copy_to(dest::MOI.ModelLike, src::MOI.ModelLike;
                  copy_names::Bool=true,
                  filter_constraints::Union{Nothing, Function}=nothing)

Use Utilities.supports_default_copy_to and Utilities.supports_allocate_load to automatically choose between Utilities.default_copy_to or Utilities.allocate_load to apply the copy operation.

If the filter_constraints arguments is given, only the constraints for which this function returns true will be copied. This function is given a constraint index as argument.

MathOptInterface.Utilities.default_copy_to — Function

default_copy_to(dest::MOI.ModelLike, src::MOI.ModelLike, copy_names::Bool,
                filter_constraints::Union{Nothing, Function}=nothing)

Implements MOI.copy_to(dest, src) by adding the variables and then the constraints and attributes incrementally. The function supports_default_copy_to can be used to check whether dest supports the copying a model incrementally.

If the filter_constraints arguments is given, only the constraints for which this function returns true will be copied. This function is given a constraint index as argument.

MathOptInterface.Utilities.supports_default_copy_to — Function

supports_default_copy_to(model::ModelLike, copy_names::Bool)

Return a Bool indicating whether the model model supports default_copy_to(model, src, copy_names=copy_names) if all the attributes set to src and constraints added to src are supported by model.

This function can be used to determine whether a model can be loaded into model incrementally or whether it should be cached and copied at once instead. This is used by JuMP to determine whether to add a cache or not in two situations:

A first cache can be used to store the model as entered by the user as well as the names of variables and constraints. This cache is created if this function returns false when copy_names is true.

If bridges are used, then a second cache can be used to store the bridged model with unnamed variables and constraints. This cache is created if this function returns false when copy_names is false.

Examples

If MathOptInterface.set, MathOptInterface.add_variable and MathOptInterface.add_constraint are implemented for a model of type MyModel and names are supported, then MathOptInterface.copy_to can be implemented as

MOI.Utilities.supports_default_copy_to(model::MyModel, copy_names::Bool) = true
function MOI.copy_to(dest::MyModel, src::MOI.ModelLike; kws...)
    return MOI.Utilities.automatic_copy_to(dest, src; kws...)
end

The Utilities.automatic_copy_to function automatically redirects to Utilities.default_copy_to.

If names are not supported, simply change the first line by

MOI.supports_default_copy_to(model::MyModel, copy_names::Bool) = !copy_names

The Utilities.default_copy_to function automatically throws an helpful error in case copy_to is called with copy_names equal to true.

Allocate-Load API

MathOptInterface.Utilities.allocate_load — Function

allocate_load(dest::MOI.ModelLike, src::MOI.ModelLike,
              filter_constraints::Union{Nothing, Function}=nothing
              )

Implements MOI.copy_to(dest, src) using the Allocate-Load API. The function supports_allocate_load can be used to check whether dest supports the Allocate-Load API.

If the filter_constraints arguments is given, only the constraints for which this function returns true will be copied. This function is given a constraint index as argument.

MathOptInterface.Utilities.supports_allocate_load — Function

supports_allocate_load(model::MOI.ModelLike, copy_names::Bool)::Bool

Return a Bool indicating whether model supports allocate_load(model, src, copy_names=copy_names) if all the attributes set to src and constraints added to src are supported by model.

MathOptInterface.Utilities.allocate_variables — Function

allocate_variables(model::MOI.ModelLike, nvars::Integer)

Creates nvars variables and returns a vector of nvars variable indices.

MathOptInterface.Utilities.allocate — Function

allocate(model::ModelLike, attr::ModelLikeAttribute, value)
allocate(model::ModelLike, attr::AbstractVariableAttribute, v::VariableIndex, value)
allocate(model::ModelLike, attr::AbstractConstraintAttribute, c::ConstraintIndex, value)

Informs model that load will be called with the same arguments after load_variables is called.

MathOptInterface.Utilities.allocate_constraint — Function

allocate_constraint(model::MOI.ModelLike, f::MOI.AbstractFunction, s::MOI.AbstractSet)

Returns the index for the constraint to be used in load_constraint that will be called after load_variables is called.

MathOptInterface.Utilities.load_variables — Function

load_variables(model::MOI.ModelLike, nvars::Integer)

Prepares model for load and load_constraint.

MathOptInterface.Utilities.load — Function

load(model::ModelLike, attr::ModelLikeAttribute, value)
load(model::ModelLike, attr::AbstractVariableAttribute, v::VariableIndex, value)
load(model::ModelLike, attr::AbstractConstraintAttribute, c::ConstraintIndex, value)

This has the same effect that set with the same arguments except that allocate should be called first before load_variables.

MathOptInterface.Utilities.load_constraint — Function

load_constraint(model::MOI.ModelLike, ci::MOI.ConstraintIndex, f::MOI.AbstractFunction, s::MOI.AbstractSet)

Sets the constraint function and set for the constraint of index ci.

Fallbacks

MathOptInterface.Utilities.get_fallback — Function

get_fallback(model::MOI.ModelLike, ::MOI.ObjectiveValue)

Compute the objective function value using the VariablePrimal results and the ObjectiveFunction value.

get_fallback(model::MOI.ModelLike, ::MOI.DualObjectiveValue, T::Type)::T

Compute the dual objective value of type T using the ConstraintDual results and the ConstraintFunction and ConstraintSet values. Note that the nonlinear part of the model is ignored.

get_fallback(model::MOI.ModelLike, ::MOI.ConstraintPrimal,
             constraint_index::MOI.ConstraintIndex)

Compute the value of the function of the constraint of index constraint_index using the VariablePrimal results and the ConstraintFunction values.

get_fallback(model::MOI.ModelLike, attr::MOI.ConstraintDual,
             ci::MOI.ConstraintIndex{Union{MOI.SingleVariable,
                                           MOI.VectorOfVariables}})

Compute the dual of the constraint of index ci using the ConstraintDual of other constraints and the ConstraintFunction values. Throws an error if some constraints are quadratic or if there is one another MOI.SingleVariable-in-S or MOI.VectorOfVariables-in-S constraint with one of the variables in the function of the constraint ci.

Function utilities

The following utilities are available for functions:

MathOptInterface.Utilities.eval_variables — Function

eval_variables(varval::Function, f::AbstractFunction)

Returns the value of function f if each variable index vi is evaluated as varval(vi). Note that varval should return a number, see substitute_variables for a similar function where varval returns a function.

MathOptInterface.Utilities.map_indices — Function

map_indices(index_map::Function, x)

Substitute any MOI.VariableIndex (resp. MOI.ConstraintIndex) in x by the MOI.VariableIndex (resp. MOI.ConstraintIndex) of the same type given by index_map(x).

This function is used by implementations of MOI.copy_to on constraint functions, attribute values and submittable values hence it needs to be implemented for custom types that are meant to be used as attribute or submittable value.

MathOptInterface.Utilities.substitute_variables — Function

substitute_variables(variable_map::Function, x)

Substitute any MOI.VariableIndex in x by variable_map(x). The variable_map function returns either MOI.SingleVariable or MOI.ScalarAffineFunction, see eval_variables for a similar function where variable_map returns a number.

This function is used by bridge optimizers on constraint functions, attribute values and submittable values when at least one variable bridge is used hence it needs to be implemented for custom types that are meant to be used as attribute or submittable value.

WARNING: Don't use substitude_variables(::Function, ...) because Julia will not specialize on this. Use instead substitude_variables(::F, ...) where {F<:Function}.

MathOptInterface.Utilities.filter_variables — Function

filter_variables(keep::Function, f::AbstractFunction)

Return a new function f with the variable vi such that !keep(vi) removed.

WARNING: Don't define filter_variables(::Function, ...) because Julia will not specialize on this. Define instead filter_variables(::F, ...) where {F<:Function}.

MathOptInterface.Utilities.remove_variable — Function

remove_variable(f::AbstractFunction, vi::VariableIndex)

Return a new function f with the variable vi removed.

remove_variable(f::MOI.AbstractFunction, s::MOI.AbstractSet, vi::MOI.VariableIndex)

Return a tuple (g, t) representing the constraint f-in-s with the variable vi removed. That is, the terms containing the variable vi in the function f are removed and the dimension of the set s is updated if needed (e.g. when f is a VectorOfVariables with vi being one of the variables).

MathOptInterface.Utilities.all_coefficients — Function

all_coefficients(p::Function, f::MOI.AbstractFunction)

Determine whether predicate p returns true for all coefficients of f, returning false as soon as the first coefficient of f for which p returns false is encountered (short-circuiting). Similar to all.

MathOptInterface.Utilities.unsafe_add — Function

unsafe_add(t1::MOI.ScalarAffineTerm, t2::MOI.ScalarAffineTerm)

Sums the coefficients of t1 and t2 and returns an output MOI.ScalarAffineTerm. It is unsafe because it uses the variable_index of t1 as the variable_index of the output without checking that it is equal to that of t2.

unsafe_add(t1::MOI.ScalarQuadraticTerm, t2::MOI.ScalarQuadraticTerm)

Sums the coefficients of t1 and t2 and returns an output MOI.ScalarQuadraticTerm. It is unsafe because it uses the variable_index's of t1 as the variable_index's of the output without checking that they are the same (up to permutation) to those of t2.

unsafe_add(t1::MOI.VectorAffineTerm, t2::MOI.VectorAffineTerm)

Sums the coefficients of t1 and t2 and returns an output MOI.VectorAffineTerm. It is unsafe because it uses the output_index and variable_index of t1 as the output_index and variable_index of the output term without checking that they are equal to those of t2.

MathOptInterface.Utilities.isapprox_zero — Function

isapprox_zero(f::MOI.AbstractFunction, tol)

Return a Bool indicating whether the function f is approximately zero using tol as a tolerance.

Important note

This function assumes that f does not contain any duplicate terms, you might want to first call canonical if that is not guaranteed. For instance, given

f = MOI.ScalarAffineFunction(MOI.ScalarAffineTerm.([1, -1], [x, x]), 0)`.

then isapprox_zero(f) is false but isapprox_zero(MOIU.canonical(f)) is true.

MathOptInterface.Utilities.modify_function — Function

modify_function(f::AbstractFunction, change::AbstractFunctionModification)

Return a new function f modified according to change.

The following functions can be used to canonicalize a function:

MathOptInterface.Utilities.is_canonical — Function

is_canonical(f::Union{ScalarAffineFunction, VectorAffineFunction})

Returns a Bool indicating whether the function is in canonical form. See canonical.

is_canonical(f::Union{ScalarQuadraticFunction, VectorQuadraticFunction})

Returns a Bool indicating whether the function is in canonical form. See canonical.

MathOptInterface.Utilities.canonical — Function

canonical(f::Union{ScalarAffineFunction, VectorAffineFunction,
                   ScalarQuadraticFunction, VectorQuadraticFunction})

Returns the function in a canonical form, i.e.

A term appear only once.

The coefficients are nonzero.

The terms appear in increasing order of variable where there the order of the variables is the order of their value.

For a AbstractVectorFunction, the terms are sorted in ascending order of output index.

The output of canonical can be assumed to be a copy of f, even for VectorOfVariables.

Examples

If x (resp. y, z) is VariableIndex(1) (resp. 2, 3). The canonical representation of ScalarAffineFunction([y, x, z, x, z], [2, 1, 3, -2, -3], 5) is ScalarAffineFunction([x, y], [-1, 2], 5).

MathOptInterface.Utilities.canonicalize! — Function

canonicalize!(f::Union{ScalarAffineFunction, VectorAffineFunction})

Convert a function to canonical form in-place, without allocating a copy to hold the result. See canonical.

canonicalize!(f::Union{ScalarQuadraticFunction, VectorQuadraticFunction})

Convert a function to canonical form in-place, without allocating a copy to hold the result. See canonical.

The following functions can be used to manipulate functions with basic algebra:

MathOptInterface.Utilities.scalar_type — Function

scalar_type(F::Type{<:MOI.AbstractVectorFunction})

Type of functions obtained by indexing objects obtained by calling eachscalar on functions of type F.

MathOptInterface.Utilities.promote_operation — Function

promote_operation(op::Function, ::Type{T},
                  ArgsTypes::Type{<:Union{T, MOI.AbstractFunction}}...) where T

Returns the type of the MOI.AbstractFunction returned to the call operate(op, T, args...) where the types of the arguments args are ArgsTypes.

MathOptInterface.Utilities.operate — Function

operate(op::Function, ::Type{T},
        args::Union{T, MOI.AbstractFunction}...)::MOI.AbstractFunction where T

Returns an MOI.AbstractFunction representing the function resulting from the operation op(args...) on functions of coefficient type T. No argument can be modified.

MathOptInterface.Utilities.operate! — Function

operate!(op::Function, ::Type{T},
         args::Union{T, MOI.AbstractFunction}...)::MOI.AbstractFunction where T

Returns an MOI.AbstractFunction representing the function resulting from the operation op(args...) on functions of coefficient type T. The first argument can be modified. The return type is the same than the method operate(op, T, args...) without !.

MathOptInterface.Utilities.operate_output_index! — Function

operate_output_index!(
    op::Function, ::Type{T}, output_index::Integer,
    func::MOI.AbstractVectorFunction
    args::Union{T, MOI.AbstractScalarFunction}...)::MOI.AbstractFunction where T

Returns an MOI.AbstractVectorFunction where the function at output_index is the result of the operation op applied to the function at output_index of func and args. The functions at output index different to output_index are the same as the functions at the same output index in func. The first argument can be modified.

MathOptInterface.Utilities.vectorize — Function

vectorize(funcs::AbstractVector{MOI.SingleVariable})

Returns the vector of scalar affine functions in the form of a MOI.VectorAffineFunction{T}.

vectorize(funcs::AbstractVector{MOI.ScalarAffineFunction{T}}) where T

Returns the vector of scalar affine functions in the form of a MOI.VectorAffineFunction{T}.

vectorize(funcs::AbstractVector{MOI.ScalarQuadraticFunction{T}}) where T

Returns the vector of scalar quadratic functions in the form of a MOI.VectorQuadraticFunction{T}.

Constraint utilities

The following utilities are available for moving the function constant to the set for scalar constraints:

MathOptInterface.Utilities.shift_constant — Function

shift_constant(set::MOI.AbstractScalarSet, offset)

Returns a new scalar set new_set such that func-in-set is equivalent to func + offset-in-new_set.

Only define this function if it makes sense to!

Use supports_shift_constant to check if the set supports shifting:

if supports_shift_constant(typeof(old_set))
    new_set = shift_constant(old_set, offset)
    f.constant = 0
    add_constraint(model, f, new_set)
else
    add_constraint(model, f, old_set)
end

See also supports_shift_constant.

Examples

The call shift_constant(MOI.Interval(-2, 3), 1) is equal to MOI.Interval(-1, 4).

MathOptInterface.Utilities.supports_shift_constant — Function

supports_shift_constant(::Type{S}) where {S<:MOI.AbstractSet}

Return true if shift_constant is defined for set S.

See also shift_constant.

MathOptInterface.Utilities.normalize_and_add_constraint — Function

normalize_and_add_constraint(model::MOI.ModelLike,
                             func::MOI.AbstractScalarFunction,
                             set::MOI.AbstractScalarSet;
                             allow_modify_function::Bool=false)

Adds the scalar constraint obtained by moving the constant term in func to the set in model. If allow_modify_function is true then the function func can be modified.

MathOptInterface.Utilities.normalize_constant — Function

normalize_constant(func::MOI.AbstractScalarFunction,
                   set::MOI.AbstractScalarSet;
                   allow_modify_function::Bool=false)

Return the func-in-set constraint in normalized form. That is, if func is MOI.ScalarQuadraticFunction or MOI.ScalarAffineFunction, the constant is moved to the set. If allow_modify_function is true then the function func can be modified.

The following utility identifies those constraints imposing bounds on a given variable, and returns those bound values:

MathOptInterface.Utilities.get_bounds — Function

get_bounds(model::MOI.ModelLike, ::Type{T}, x::MOI.VariableIndex)

Return a tuple (lb, ub) of type Tuple{T, T}, where lb and ub are lower and upper bounds, respectively, imposed on x in model.

The following utilities are useful when working with symmetric matrix cones.

MathOptInterface.Utilities.is_diagonal_vectorized_index — Function

is_diagonal_vectorized_index(index::Base.Integer)

Return whether index is the index of a diagonal element in a MOI.AbstractSymmetricMatrixSetTriangle set.

MathOptInterface.Utilities.side_dimension_for_vectorized_dimension — Function

side_dimension_for_vectorized_dimension(n::Integer)

Return the dimension d such that MOI.dimension(MOI.PositiveSemidefiniteConeTriangle(d)) is n.