Containers API

More information can be found in the Containers section of the manual.

JuMP.Containers — Module

Containers

Module defining the containers DenseAxisArray and SparseAxisArray that behaves as a regular AbstractArray but with custom indexes that are not necessarily integers.