Model API

More information can be found in the Interacting with solvers section of the manual.

JuMP.Model — Method

Model(; caching_mode::MOIU.CachingOptimizerMode=MOIU.AUTOMATIC)

Return a new JuMP model without any optimizer; the model is stored the model in a cache. The mode of the CachingOptimizer storing this cache is caching_mode. Use set_optimizer to set the optimizer before calling optimize!.

JuMP.Model — Method

Model(optimizer_factory;
      caching_mode::MOIU.CachingOptimizerMode=MOIU.AUTOMATIC,
      bridge_constraints::Bool=true)

Return a new JuMP model with the provided optimizer and bridge settings. This function is equivalent to:

    model = Model()
    set_optimizer(model, optimizer_factory,
                  bridge_constraints=bridge_constraints)
    return model

See set_optimizer for the description of the optimizer_factory and bridge_constraints arguments.

Examples

The following creates a model with the optimizer set to Ipopt:

model = Model(Ipopt.Optimizer)

JuMP.direct_model — Function

direct_model(backend::MOI.ModelLike)

Return a new JuMP model using backend to store the model and solve it. As opposed to the Model constructor, no cache of the model is stored outside of backend and no bridges are automatically applied to backend. The absence of cache reduces the memory footprint but it is important to bear in mind the following implications of creating models using this direct mode:

When backend does not support an operation, such as modifying constraints or adding variables/constraints after solving, an error is thrown. For models created using the Model constructor, such situations can be dealt with by storing the modifications in a cache and loading them into the optimizer when optimize! is called.

No constraint bridging is supported by default.

The optimizer used cannot be changed the model is constructed.

The model created cannot be copied.

JuMP.set_optimizer — Function

set_optimizer(model::Model, optimizer_factory;
              bridge_constraints::Bool=true)

Creates an empty MathOptInterface.AbstractOptimizer instance by calling optimizer_factory() and sets it as the optimizer of model. Specifically, optimizer_factory must be callable with zero arguments and return an empty MathOptInterface.AbstractOptimizer.

If bridge_constraints is true, constraints that are not supported by the optimizer are automatically bridged to equivalent supported constraints when an appropriate transformation is defined in the MathOptInterface.Bridges module or is defined in another module and is explicitly added.

See set_optimizer_attributes and set_optimizer_attribute for setting solver-specific parameters of the optimizer.

Examples

model = Model()
set_optimizer(model, GLPK.Optimizer)

JuMP.optimizer_with_attributes — Function

optimizer_with_attributes(optimizer_constructor, attrs::Pair...)

Groups an optimizer constructor with the list of attributes attrs. Note that it is equivalent to MOI.OptimizerWithAttributes.

When provided to the Model constructor or to set_optimizer, it creates an optimizer by calling optimizer_constructor(), and then sets the attributes using set_optimizer_attribute.

Example

model = Model(
    optimizer_with_attributes(
        Gurobi.Optimizer, "Presolve" => 0, "OutputFlag" => 1
    )
)

is equivalent to:

model = Model(Gurobi.Optimizer)
set_optimizer_attribute(model, "Presolve", 0)
set_optimizer_attribute(model, "OutputFlag", 1)

Note

The string names of the attributes are specific to each solver. One should consult the solver's documentation to find the attributes of interest.

See also: set_optimizer_attribute, set_optimizer_attributes, get_optimizer_attribute.

JuMP.set_silent — Function

set_silent(model::Model)

Takes precedence over any other attribute controlling verbosity and requires the solver to produce no output.

JuMP.unset_silent — Function

unset_silent(model::Model)

Neutralize the effect of the set_silent function and let the solver attributes control the verbosity.

JuMP.set_time_limit_sec — Function

set_time_limit_sec(model::Model, limit)

Sets the time limit (in seconds) of the solver. Can be unset using unset_time_limit_sec or with limit set to nothing.

JuMP.unset_time_limit_sec — Function

unset_time_limit_sec(model::Model)

Unsets the time limit of the solver. Can be set using set_time_limit_sec.

JuMP.time_limit_sec — Function

time_limit_sec(model::Model)

Gets the time limit (in seconds) of the model (nothing if unset). Can be set using set_time_limit_sec.

JuMP.solver_name — Function

solver_name(model::Model)

If available, returns the SolverName property of the underlying optimizer. Returns "No optimizer attached" in AUTOMATIC or MANUAL modes when no optimizer is attached. Returns "SolverName() attribute not implemented by the optimizer." if the attribute is not implemented.

JuMP.backend — Function

backend(model::Model)

Return the lower-level MathOptInterface model that sits underneath JuMP. This model depends on which operating mode JuMP is in (manual, automatic, or direct), and whether there are any bridges in the model.

If JuMP is in direct mode (i.e., the model was created using direct_model), the backend with be the optimizer passed to direct_model. If JuMP is in manual or automatic mode, the backend is a MOI.Utilities.CachingOptimizer.

This function should only be used by advanced users looking to access low-level MathOptInterface or solver-specific functionality.

Base.empty! — Method

empty!(model::Model) -> model

Empty the model, that is, remove all variables, constraints and model attributes but not optimizer attributes. Always return the argument.

Note: removes extensions data.

JuMP.get_optimizer_attribute — Function

get_optimizer_attribute(model, name::String)

Return the value associated with the solver-specific attribute named name.

Note that this is equivalent to get_optimizer_attribute(model, MOI.RawParameter(name)).

Example

get_optimizer_attribute(model, "SolverSpecificAttributeName")

See also: set_optimizer_attribute, set_optimizer_attributes.

get_optimizer_attribute(
    model::Model, attr::MOI.AbstractOptimizerAttribute
)

Return the value of the solver-specific attribute attr in model.

Example

get_optimizer_attribute(model, MOI.Silent())

See also: set_optimizer_attribute, set_optimizer_attributes.

JuMP.set_optimizer_attribute — Function

set_optimizer_attribute(model::Model, name::String, value)

Sets solver-specific attribute identified by name to value.

Note that this is equivalent to set_optimizer_attribute(model, MOI.RawParameter(name), value).

Example

set_optimizer_attribute(model, "SolverSpecificAttributeName", true)

See also: set_optimizer_attributes, get_optimizer_attribute.

set_optimizer_attribute(
    model::Model, attr::MOI.AbstractOptimizerAttribute, value
)

Set the solver-specific attribute attr in model to value.

Example

set_optimizer_attribute(model, MOI.Silent(), true)

See also: set_optimizer_attributes, get_optimizer_attribute.

JuMP.set_optimizer_attributes — Function

set_optimizer_attributes(model::Model, pairs::Pair...)

Given a list of attribute => value pairs, calls set_optimizer_attribute(model, attribute, value) for each pair.

Example

model = Model(Ipopt.Optimizer)
set_optimizer_attributes(model, "tol" => 1e-4, "max_iter" => 100)

is equivalent to:

model = Model(Ipopt.Optimizer)
set_optimizer_attribute(model, "tol", 1e-4)
set_optimizer_attribute(model, "max_iter", 100)

See also: set_optimizer_attribute, get_optimizer_attribute.

JuMP.bridge_constraints — Function

bridge_constraints(model::Model)

When in direct mode, return false. When in manual or automatic mode, return a Bool indicating whether the optimizer is set and unsupported constraints are automatically bridged to equivalent supported constraints when an appropriate transformation is available.

JuMP.print_bridge_graph — Function

 print_bridge_graph([io::IO,] model::Model)

Print the hyper-graph containing all variable, constraint, and objective types that could be obtained by bridging the variables, constraints, and objectives that are present in the model.

Each node in the hyper-graph corresponds to a variable, constraint, or objective type.

Variable nodes are indicated by [ ]

Constraint nodes are indicated by ( )

Objective nodes are indicated by | |

The number inside each pair of brackets is an index of the node in the hyper-graph.

Note that this hyper-graph is the full list of possible transformations. When the bridged model is created, we select the shortest hyper-path(s) from this graph, so many nodes may be un-used.

For more information, see Legat, B., Dowson, O., Garcia, J., and Lubin, M. (2020). "MathOptInterface: a data structure for mathematical optimization problems." URL: https://arxiv.org/abs/2002.03447

JuMP.write_to_file — Function

write_to_file(
    model::Model,
    filename::String;
    format::MOI.FileFormats.FileFormat = MOI.FileFormats.FORMAT_AUTOMATIC
)

Write the JuMP model model to filename in the format format.

If the filename ends in .gz, it will be compressed using Gzip. If the filename ends in .bz2, it will be compressed using BZip2.

Base.write — Method

Base.write(
    io::IO,
    model::Model;
    format::MOI.FileFormats.FileFormat = MOI.FileFormats.FORMAT_MOF
)

Write the JuMP model model to io in the format format.

JuMP.read_from_file — Function

read_from_file(
    filename::String;
    format::MOI.FileFormats.FileFormat = MOI.FileFormats.FORMAT_AUTOMATIC
)

Return a JuMP model read from filename in the format format.

If the filename ends in .gz, it will be uncompressed using Gzip. If the filename ends in .bz2, it will be uncompressed using BZip2.

Base.read — Method

Base.read(io::IO, ::Type{Model}; format::MOI.FileFormats.FileFormat)

Return a JuMP model read from io in the format format.