Nonlinear Modeling

More information can be found in the Nonlinear section of the manual.

JuMP.Nonlinear.NonlinearData — Type

NonlinearData()

Construct a new MOI.AbstractNLPEvaluator object.

Expressions

JuMP.Nonlinear.ExpressionIndex — Type

ExpressionIndex

JuMP.Nonlinear.add_expression — Function

add_expression(data::NonlinearData, expr)::ExpressionIndex

Parse expr into a NonlinearExpression and add to data. Returns an ExpressionIndex that can be interpolated into other input expressions.

Examples

data = NonlinearData()
x = MOI.VariableIndex(1)
ex = add_expression(data, :($x^2 + 1))
set_objective(data, :(sqrt($ex)))

Parameters

JuMP.Nonlinear.ParameterIndex — Type

ParameterIndex

JuMP.Nonlinear.add_parameter — Function

add_parameter(data::NonlinearData, value::Float64)::ParameterIndex

Add a new parameter to data with the default value value. Returns a ParameterIndex that can be interpolated into other input expressions and used to modify the value of the parameter.

Examples

data = NonlinearData()
x = MOI.VariableIndex(1)
p = add_parameter(data, 1.2)
c = add_constraint(data, :($x^2 <= $p))

Objectives

JuMP.Nonlinear.set_objective — Function

set_objective(data::NonlinearData, obj)::Nothing

Parse obj into a NonlinearExpression and set as the objective function of data.

Examples

data = NonlinearData()
x = MOI.VariableIndex(1)
set_objective(data, :($x^2 + 1))
set_objective(data, x)

Constraints

JuMP.Nonlinear.ConstraintIndex — Type

ConstraintIndex

JuMP.Nonlinear.add_constraint — Function

add_constraint(data::NonlinearData, input::Expr)::ConstraintIndex

Parse input into a NonlinearConstraint and add to data. Returns an ConstraintIndex that can be used to delete the constraint or query solution information.

Examples

data = NonlinearData()
x = MOI.VariableIndex(1)
c = add_constraint(data, :($x^2 <= 1))

JuMP.Nonlinear.delete — Function

delete(data::NonlinearData, c::ConstraintIndex)::Nothing

Delete the constraint index c from data.

Examples

data = NonlinearData()
x = MOI.VariableIndex(1)
c = add_constraint(data, :($x^2 <= 1))
delete(data, c)

User-defined operators

JuMP.Nonlinear.OperatorRegistry — Type

OperatorRegistry()

Create a new OperatorRegistry to store and evaluate univariate and multivariate operators.

JuMP.Nonlinear.DEFAULT_UNIVARIATE_OPERATORS — Constant

DEFAULT_UNIVARIATE_OPERATORS

The list of univariate operators that are supported by default.

JuMP.Nonlinear.DEFAULT_MULTIVARIATE_OPERATORS — Constant

DEFAULT_MULTIVARIATE_OPERATORS

The list of multivariate operators that are supported by default.

JuMP.Nonlinear.register_operator — Function

register_operator(
    data::NonlinearData,
    op::Symbol,
    nargs::Int,
    f::Function,
    [∇f::Function],
    [∇²f::Function],
)

Register the user-defined operator op with nargs input arguments in data.

Univariate functions

f(x::T)::T must be a function that takes a single input argument x and returns the function evaluated at x. If ∇f and ∇²f are not provided, f must support any Real input type T.

∇f(x::T)::T is a function that takes a single input argument x and returns the first derivative of f with respect to x. If ∇²f is not provided, ∇f must support any Real input type T.

∇²f(x::T)::T is a function that takes a single input argument x and returns the second derivative of f with respect to x.

Multivariate functions

f(x::T...)::T must be a function that takes a nargs input arguments x and returns the function evaluated at x. If ∇f and ∇²f are not provided, f must support any Real input type T.

∇f(g::AbstractVector{T}, x::T...)::T is a function that takes a cache vector g of length length(x), and fills each element g[i] with the partial derivative of f with respect to x[i].

Hessian are not supported for multivariate functions.

JuMP.Nonlinear.register_operator_if_needed — Function

register_operator_if_needed(
    registry::OperatorRegistry,
    op::Symbol,
    nargs::Int,
    f::Function;
)

Similar to register_operator, but this function warns if the function is not registered, and skips silently if it already is.

JuMP.Nonlinear.assert_registered — Function

assert_registered(registry::OperatorRegistry, op::Symbol, nargs::Int)

Throw an error if op is not registered in registry with nargs arguments.

JuMP.Nonlinear.check_return_type — Function

check_return_type(::Type{T}, ret::S) where {T,S}

Overload this method for new types S to throw an informative error if a user-defined function returns the type S instead of T.

JuMP.Nonlinear.eval_univariate_function — Function

eval_univariate_function(
    registry::OperatorRegistry,
    op::Symbol,
    x::T,
) where {T}

Evaluate the operator op(x)::T, where op is a univariate function in registry.

JuMP.Nonlinear.eval_univariate_gradient — Function

eval_univariate_gradient(
    registry::OperatorRegistry,
    op::Symbol,
    x::T,
) where {T}

Evaluate the first-derivative of the operator op(x)::T, where op is a univariate function in registry.

JuMP.Nonlinear.eval_univariate_hessian — Function

eval_univariate_hessian(
    registry::OperatorRegistry,
    op::Symbol,
    x::T,
) where {T}

Evaluate the second-derivative of the operator op(x)::T, where op is a univariate function in registry.

JuMP.Nonlinear.eval_multivariate_function — Function

eval_multivariate_function(
    registry::OperatorRegistry,
    op::Symbol,
    x::AbstractVector{T},
) where {T}

Evaluate the operator op(x)::T, where op is a multivariate function in registry.

JuMP.Nonlinear.eval_multivariate_gradient — Function

eval_multivariate_gradient(
    registry::OperatorRegistry,
    op::Symbol,
    g::AbstractVector{T},
    x::AbstractVector{T},
) where {T}

Evaluate the gradient of operator g .= ∇op(x), where op is a multivariate function in registry.

JuMP.Nonlinear.eval_logic_function — Function

eval_logic_function(
    registry::OperatorRegistry,
    op::Symbol,
    lhs::T,
    rhs::T,
)::Bool where {T}

Evaluate (lhs op rhs)::Bool, where op is a logic operator in registry.

JuMP.Nonlinear.eval_comparison_function — Function

eval_comparison_function(
    registry::OperatorRegistry,
    op::Symbol,
    lhs::T,
    rhs::T,
)::Bool where {T}

Evaluate (lhs op rhs)::Bool, where op is a comparison operator in registry.

Automatic-differentiation backends

JuMP.Nonlinear.AbstractAutomaticDifferentiation — Type

AbstractAutomaticDifferentiation

An abstract type for extending

JuMP.Nonlinear.Default — Type

Default() <: AbstractAutomaticDifferentiation

The default implementation of AbstractAutomaticDifferentiation. This only supported :ExprGraph.

JuMP.Nonlinear.SparseReverseMode — Type

SparseReverseMode() <: AbstractAutomaticDifferentiation

JuMP.Nonlinear.set_differentiation_backend — Function

set_differentiation_backend(
    data::NonlinearData,
    backend::AbstractAutomaticDifferentiation,
    ordered_variables::Vector{MOI.VariableIndex}
)

Set the automatic differentiationn backend of data to backend.

ordered_variables is an ordered vector containing all of the variables that are present in the model. This order corresponds to the order of the primal solution vector x that is passed to the various functions in MOI's nonlinear API.

Data-structure

JuMP.Nonlinear.Node — Type

struct Node
    type::NodeType
    index::Int
    parent::Int
end

JuMP.Nonlinear.NodeType — Type

NodeType

An enum describing the possible node types. Each Node has a .index field, which should be interpreted as follows:

NODE_CALL_MULTIVARIATE: the index into operators.multivariate_operators

NODE_CALL_UNIVARIATE: the index into operators.univariate_operators

NODE_LOGIC: the index into operators.logic_operators

NODE_COMPARISON: the index into operators.comparison_operators

NODE_MOI_VARIABLE: the value of MOI.VariableIndex(index) in the user's space of the model.

NODE_VARIABLE: the 1-based index of the internal vector

NODE_VALUE: the index into the .values field of NonlinearExpression

NODE_PARAMETER: the index into data.parameters

NODE_SUBEXPRESSION: the index into data.expressions

JuMP.Nonlinear.NonlinearExpression — Type

struct NonlinearExpression
    nodes::Vector{Node}
    values::Vector{Float64}
end

JuMP.Nonlinear.adjacency_matrix — Function

adjacency_matrix(nodes::Vector{Node})

Compute the sparse adjacency matrix describing the parent-child relationships in nodes.

The element (i, j) is true if there is an edge from node[j] to node[i]. Since we get a column-oriented matrix, this gives us a fast way to look up the edges leaving any node (i.e., the children).